Move some tests from org.enso.interpreter.test to dedicated packages
