+++ engine/runtime-integration-tests/src/test/java/org/enso/interpreter/test/instrument/InsightForEnsoTest.java
@@ -1,4 +1,4 @@
-package org.enso.interpreter.test;
+package org.enso.interpreter.test.instrument;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
