+++ engine/runtime-integration-tests/src/test/java/org/enso/interpreter/test/interop/JsInteropTest.java
@@ -1,4 +1,4 @@
-package org.enso.interpreter.test;
+package org.enso.interpreter.test.interop;
 
 import static org.junit.Assert.assertEquals;
