+++ engine/runtime-integration-tests/src/test/java/org/enso/interpreter/test/meta/MetaIsAPolyglotTest.java
@@ -1,5 +1,6 @@
-package org.enso.interpreter.test;
+package org.enso.interpreter.test.meta;
 
+import org.enso.interpreter.test.ValuesGenerator;
 import org.enso.interpreter.test.ValuesGenerator.Language;
 import org.graalvm.polyglot.Context;
